Add validation for duplicate state names
