+++ tests/test_pass_state.py
@@ -58,19 +58,3 @@ Running Pass 3
 Passing
 """
     )
-
-
-def test_one_state(compile_state_machine):
-    pass_state = PassState("My Pass", comment="The only state")
-    state_machine = StateMachine(start_state=pass_state)
-    compiled = compile_state_machine(state_machine)
-    assert compiled == {
-        "StartAt": pass_state.name,
-        "States": {
-            pass_state.name: {
-                "Comment": pass_state.comment,
-                "Type": "Pass",
-                "End": True,
-            },
-        },
-    }
